requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pdfminer.six>=20221105
python-docx>=1.1.0
spacy>=3.7.0
scikit-learn>=1.3.0
pandas>=2.1.0
huggingface-hub>=0.19.0
openai>=1.3.0

